+++ Ryujinx.Graphics.Texture/BC6Decoder.cs
@@ -99,7 +99,7 @@ namespace Ryujinx.Graphics.Texture
                     RgbaColor32 color = BC67Utils.Interpolate(color1, color2, index, indexBitCount);
 
                     output[offs] =
-                        (ulong)FinishUnquantize(color.R, signed) |
+                        FinishUnquantize(color.R, signed) |
                         ((ulong)FinishUnquantize(color.G, signed) << 16) |
                         ((ulong)FinishUnquantize(color.B, signed) << 32) |
                         ((ulong)HalfOne << 48);
